--- a/changelogs/unreleased/gh-11404-speed-up-txn-limbo-confirm-writing.md
+++ b/changelogs/unreleased/gh-11404-speed-up-txn-limbo-confirm-writing.md
@@ -1,0 +1,4 @@
+## bugfix/limbo
+
+* Optimized limbo worker writing confirmations for synchronous transactions
+  (gh-11404).