In the case of a single node cluster, the limbo worker was not writing synchro confirms fast enough.
* The problem was, firstly, that sometimes it was possible to queue a confirm writing immediately (without yielding) from the `txn_on_journal_write` write-callback, instead of waking up the limbo worker to queue this writing on the next iteration of the event loop.
* The second problem was that limbo worker wrote confirms synchronously. There is no point in waiting for the confirm writing to finish, because while the worker is waiting, `volatile_confirmed_lsn` could have bumped and it doesn't matter how the previous writing ends, because you can already write a confirm to the updated greater lsn and retry only it if necessary, there is no point in retry unsuccessful writes with lower lsn.

```console
tarantool:~$ tarantool perf/lua/1mops_write.lua --nodes=31 --fibers=6000 --ops=1000000 --transaction=1 --warmup=10 --sync
# making 1000000 REPLACE operations,
# 1 operations per txn,
# using 6000 fibers,
...
```

<table border="1" cellspacing="0" cellpadding="8">
    <thead>
        <tr>
            <th>--nodes=</th>
            <th>Before</th>
            <th>After</th>
        </tr>
    </thead>
    <tbody>
        <tr>
            <td>1</td>
            <td>master average speed: 283773 ops/sec</td>
            <td>master average speed: 586668 ops/sec</td>
        </tr>
        <tr>
            <td rowspan="2">2</td>
            <td>master average speed: 233323 ops/sec</td>
            <td>master average speed: 233565 ops/sec</td>
        </tr>
        <tr>
            <td>1mops_replica_rps: 233320</td>
            <td>1mops_replica_rps: 233563</td>
        </tr>
        <tr>
            <td rowspan="2">3</td>
            <td>master average speed: 196305 ops/sec</td>
            <td>master average speed: 205243 ops/sec</td>
        </tr>
        <tr>
            <td>1mops_replica_rps: 196303</td>
            <td>1mops_replica_rps: 205235</td>
        </tr>
        <tr>
            <td rowspan="2">31</td>
            <td>master average speed: 49210 ops/sec</td>
            <td>master average speed: 49945 ops/sec</td>
        </tr>
        <tr>
            <td>1mops_replica_rps: 49207</td>
            <td>1mops_replica_rps: 49943</td>
        </tr>
    </tbody>
</table>

Closes https://github.com/tarantool/tarantool/issues/11404

NO_DOC=performance improvement
NO_TEST=performance improvement